\name{DPQ-package}
\alias{DPQ-package}
\alias{DPQ}
\docType{package}
\title{
\packageTitle{DPQ}
}
\description{
\packageDescription{DPQ}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{DPQ}
\packageIndices{DPQ}
%% ~~ An overview of how to use the package, including the most important ~~
%% ~~ functions ~~
}
\author{
\packageAuthor{DPQ}

Maintainer: \packageMaintainer{DPQ}
}
%% \references{
%% ~~ Literature or other references for background information ~~
%% }
\seealso{
  The package \code{\link[DPQmpfr:DPQmpfr-package]{DPQmpfr}}, which
  builds on this package and on
  \code{\link[Rmpfr:Rmpfr-package]{Rmpfr}}.
}
%% \examples{
TODO
%% }
\keyword{ distribution }
\keyword{ math }
\keyword{ package }
