\name{DPQ-package}
\alias{DPQ-package}
\alias{DPQ}
\docType{package}
\title{\packageTitle{DPQ}}
\description{
  \packageDescription{DPQ}
}
\details{
  The DESCRIPTION file:
  \packageDESCRIPTION{DPQ}
  \packageIndices{DPQ}
  %% ~~ Overview of how to use the package, incl. the most important functions ~~
}
\author{
  Principal author and maintainer: \packageMaintainer{DPQ}
  %% \packageAuthor{DPQ}
}
%% \references{
%% ~~ Literature or other references for background information ~~
%% }
\seealso{
  The package %\pkg{\link[DPQmpfr:DPQmpfr-package]{DPQmpfr}},
  \pkg{DPQmpfr} (not yet on CRAN),
  which builds on this package and on
  \pkg{\link[Rmpfr:Rmpfr-package]{Rmpfr}}.
}
\examples{
"FIXME"
}
\keyword{ distribution }
\keyword{ math }
\keyword{ package }
